# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My User Project CRUD",
    "description": "Users, products and orders APIs",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Products",
      "description": "API for products in the system"
    },
    {
      "name": "Orders",
      "description": "API for orders in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Find all products",
        "description": "Returns all products.",
        "responses": {
          "200": {
            "description": "Product document found"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/products/{id}": {
      "delete": {
        "tags": ["Products"],
        "summary": "Delete product by id",
        "description": "Delete product by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product document deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  }
}